requirements.txt
absl-py==0.14.1
appnope==0.1.2
astunparse==1.6.3
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
black==21.9b0
cachelib==0.4.1
cachetools==4.2.4
certifi==2021.5.30
chardet==3.0.4
clang==5.0
click==8.0.1
cycler==0.10.0
Cython==0.29.24
dartsclone==0.9.0
debugpy==1.4.3
decorator==5.1.0
emoji==1.5.0
entrypoints==0.3
filelock==3.3.0
Flask==2.0.2
Flask-Session==0.4.0
flatbuffers==1.12
fugashi==1.1.1
funcy==1.16
future==0.18.2
gast==0.4.0
gensim==4.1.2
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==3.0.0
grpcio==1.41.0
h11==0.9.0
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
hstspreload==2021.9.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.0.19
hyperframe==5.2.0
idna==2.10
ipadic==1.0.0
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jupyter-client==7.0.3
jupyter-core==4.8.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mecab-python3==1.0.4
mypy-extensions==0.4.3
neologdn==0.5.1
nest-asyncio==1.5.1
nltk==3.6.3
numexpr==2.7.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.3.0
prompt-toolkit==3.0.20
protobuf==3.18.1
psycopg==3.0b1
psycopg2==2.9.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.10.0
pyLDAvis==3.3.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.3.0
regex==2021.8.28
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.7.2
sacremoses==0.0.46
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
searchtweets-v2==1.1.1
six==1.15.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.1.0
soupsieve==2.2.1
SudachiDict-core==20210802
SudachiPy==0.5.3
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==2.2.0
tokenizers==0.10.3
tomli==1.2.1
torch==1.9.1
torchaudio==0.9.1
torchvision==0.10.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
transformers==4.11.3
tweepy==3.10.0
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.12.1
